% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_stuff.R
\name{best_wflow_preds_vi}
\alias{best_wflow_preds_vi}
\title{Get "best" parameter sets for each modeling approach (RF, DT, GLM-EN), and also generating metrics of fit on training data}
\usage{
best_wflow_preds_vi(
  regression_workflow_results,
  outcome_var,
  classification = FALSE,
  metric = "rmse",
  train_sample = train,
  test_sample = test,
  vi = TRUE
)
}
\arguments{
\item{classification}{is a logical to tell us if this is a classification model. in which case we want the ROC curve stuff}

\item{metric}{is our measure of fit, I guess, like rmse or roc_auc}

\item{vi}{if we want variable importance measures}
}
\description{
Get "best" parameter sets for each modeling approach (RF, DT, GLM-EN), and also generating metrics of fit on training data
}
\details{
returns a tibble of the workflow ('output' is a 'fit workflow') with the best parameters, predictions, true outcomes, and some metrics
}
\examples{
Used in best_wflow_preds_vi, also...
best_dt_preproc <- best_wflow_preds_vi(preproc_results, metric = 'roc_auc', classification = TRUE,
                                       outcome_var = "showup",
                                       train_sample = eas_all_train,
                                       test_sample = eas_all_test)
}
