% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.r
\name{book}
\alias{book}
\title{The Rethink Priorities bookdown output format}
\usage{
book(
  toc = TRUE,
  number_sections = TRUE,
  fig_caption = TRUE,
  lib_dir = "libs",
  global_numbering = !number_sections,
  pandoc_args = NULL,
  ...,
  base_format = rmarkdown::html_document,
  split_bib = TRUE,
  split_by = c("section+number", "section", "chapter+number", "chapter", "rmd", "none")
)
}
\description{
The Rethink Priorities bookdown output format
}
