% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sink.R
\name{clean_sink}
\alias{clean_sink}
\title{Puts together a bunch of common 'recipe' steps used in ML and other contexts}
\usage{
clean_sink(df)
}
\description{
Puts together a bunch of common 'recipe' steps used in ML and other contexts
}
\details{
Not used in EAS code
}
