% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_one_hot_back.R
\name{convert_one_hot_back}
\alias{convert_one_hot_back}
\title{Converts 'one-hot' Dummies to factors}
\usage{
convert_one_hot_back(df, prefix, vars, keep_dummies = FALSE)
}
\arguments{
\item{XXX}{describe an input to function}
}
\description{
\code{convert_one_hot_back} is a function that...
}
\details{
I think I got it from https://stackoverflow.com/questions/64230674/how-to-turn-one-hot-encoded-variables-to-a-single-factor-in-r
}
\examples{
  dat <- data.frame(season = as.factor(sample(1:4, 100, replace = TRUE))) 
  dat2 <- model.matrix(~ season, dat)
  convert_one_hot_back(as_tibble(dat2), "season", c("season2", "season3", "season4")) \%>\% head()
}
