% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_hijacks.R
\name{.kable_styling}
\alias{.kable_styling}
\title{Convenience 'hijacks' of kable and kable_styling to get formats I like}
\usage{
.kable_styling(
  kable_input,
  bootstrap_options = "basic",
  latex_options = "basic",
  full_width = FALSE,
  position = "center",
  font_size = NULL,
  row_label_position = "l",
  repeat_header_text = "\\\\textit{(continued)}",
  repeat_header_method = c("append", "replace"),
  repeat_header_continued = FALSE,
  stripe_color = "gray!6",
  stripe_index = NULL,
  latex_table_env = NULL,
  protect_latex = TRUE,
  table.envir = "table",
  fixed_thead = FALSE,
  htmltable_class = NULL,
  html_font = NULL,
  wraptable_width = "0pt"
)
}
\description{
\code{.kable_styling} ma
}
\examples{
  mtcars \%>\%
     .kable(digits=2) \%>\%
    .kable_styling()
}
