% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coefficients.R
\name{extract_coefficients}
\alias{extract_coefficients}
\title{Make a model into a tidy object}
\usage{
extract_coefficients(
  fit,
  replacement_names = NULL,
  model_name = NULL,
  robust_SE = TRUE,
  exponentiate = FALSE
)
}
\arguments{
\item{XXX}{describe an input to function}
}
\description{
\code{extract_coefficients} is a function that extracts (and rename) coefficients and CIs (from a list/tibble including columns 'term' and ??)
}
\details{
Need to add more details here. Note that robust sandwich se's are defaults
}
