% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_test_helpers_infer.R
\name{extract_hyp_results}
\alias{extract_hyp_results}
\title{\code{extract_hyp_results} is a function to extract key results from hypothesis testing using the infer package}
\usage{
extract_hyp_results(df, remove = c("df"))
}
\description{
\code{extract_hyp_results} is a function to extract key results from hypothesis testing using the infer package
}
\details{
Don't run for independence (this doesn't specify a statistic value) ...
Unnests results
Changes name of mu, med, p to "null_value"
Displays formulas in text form
}
\examples{
See 'donations_sparse.Rmd'
linked_tests_results <- extract_hyp_results(linked_tests_df) \%>\% # extract and label key results for reporting and plotting mutat
(data_label = linked_df_labels,
    data_type = linked_test_var_type)
    
}
