% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_test_helpers_infer.R
\name{extract_hyp_value}
\alias{extract_hyp_value}
\title{Hypothesis test helpers for working with \code{infer} package
\code{extract_hyp_value} is a function to switch hypothesis statistic names to value for standardized output. It is mainly used in \code{extract_hyp_results}.}
\usage{
extract_hyp_value(hypothesis)
}
\description{
Hypothesis test helpers for working with \code{infer} package
\code{extract_hyp_value} is a function to switch hypothesis statistic names to value for standardized output. It is mainly used in \code{extract_hyp_results}.
}
\details{
Don't run for independence (this doesn't specify a statistic value)
}
