% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit linear (regression) model and return fit (not tidied, just regular)}
\usage{
fit_glm(
  formula,
  df,
  family = quasipoisson,
  replacement_names = NULL,
  keep_data = TRUE,
  y = TRUE,
  qr = TRUE,
  keep_model_matrix = FALSE
)
}
\arguments{
\item{keep_data}{whether to return the model frame?}

\item{qr}{whether to keep the 'QR decomposition', a factorization of the (?) model matrix}

\item{keep_model_matrix}{whether to keep the model matrix}
}
\description{
\code{fit_glm} is like fit_lm but for a glm I guess ...  a function that fits a general linear regression model and returns a 'model frame', and can keep the model matrix, etc.
}
\examples{
 linear_models <- linear_models \%>\%
mutate(
 lm_fit = fit_models(
   linear_models, "formulas", "dfs", fun = fit_lm)
)
}
