% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit models using columns specifying formulas and required dataframes;}
\usage{
fit_models(df, formulas_col, dfs_col, fun = fit_lm, family = NULL)
}
\description{
Fit models using columns specifying formulas and required dataframes;
}
\details{
fit_lm as default modeling thing, but fit_glm, etc can also work
}
\examples{
# Create data frame for modeling
linear_models <- make_model_df(rhs_vars_list, outcome_vars_list, dfs)
Fit linear models
linear_models <- linear_models \%>\%
  mutate(
    lm_fit = fit_models(
    linear_models, "formulas", "dfs", fun = fit_lm)
  )
  
}
