% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_model_stuff.R
\name{get_best_result}
\alias{get_best_result}
\title{Extract best parameters etc from tidymodels workflow}
\usage{
get_best_result(results_df, id, metric = "rmse")
}
\description{
Extract best parameters etc from tidymodels workflow
}
\details{
Given a workflow extract the best parameters and  the corresponding workflow
}
\examples{
 Used in best_wflow_preds_vi
}
