% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_tidy_helpers.R, R/tidy_model_stuff.R
\name{get_var_importance}
\alias{get_var_importance}
\title{Machine learning Tidymodels helpers}
\usage{
get_var_importance(fit)

get_var_importance(fit)
}
\arguments{
\item{fit}{is a tune_results object, I think}
}
\description{
\code{get_var_importance} is a function to get model fit and make a variable importance plot, I guess
\code{scale_var} is just a code shortcut for scaling
\code{norm_vi} prepares content for importance plots
\code{plot_vi} creqtes varln importance plots
}
\details{
\code{get_var_importance}  needs workflowsets and vip packages, part of the tidymodels ecosystem
}
\examples{
 best_models$fit[[1]] \%>\% get_var_importance()
}
