% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_sum.R
\name{grp_sum}
\alias{grp_sum}
\title{Grouped summary table of 2 variables, mainly for plotting bivariate means by groups
\code{grp_sum} creates a table of summary statistics on two variables, grouped by a third variable.}
\usage{
grp_sum(df, xvar, yvar, groupvar)
}
\arguments{
\item{XXX}{describe an input to function}
}
\description{
Grouped summary table of 2 variables, mainly for plotting bivariate means by groups
\code{grp_sum} creates a table of summary statistics on two variables, grouped by a third variable.
}
\details{

}
\examples{
 XXX nothing(mtcars)
}
