% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hijack.R
\name{hijack}
\alias{hijack}
\title{Shortcut tool for altering a function}
\usage{
hijack(FUN, ...)
}
\arguments{
\item{FUN}{An already defined function}
}
\description{
Shortcut tool for altering a function
}
\details{
Taken from  https://www.r-bloggers.com/hijacking-r-functions-changing-default-arguments/
}
\examples{
# Here is an example
.median <- hijack(stats::median, na.rm = TRUE)

}
