% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_test_infer_sim.R
\name{htest_infer_sim}
\alias{htest_infer_sim}
\title{helper for randomization-inference simulation testing using Infer package; set options in args}
\usage{
htest_infer_sim(
  df,
  formula,
  response = NULL,
  hypothesis,
  gen,
  stat = NULL,
  p_val_dir = NULL,
  seed = 123
)
}
\arguments{
\item{formula}{I'm not completely sure how to use a formula here, does it transform a variable?}

\item{response}{the outcome of interest}

\item{hypothesis}{\itemize{
\item If the hypothesis option stat is specified then statistics like the mean are calculated for this sample
}}

\item{gen}{something about how many samples are generated and how}

\item{stat}{the statistic(s) (multiple?) to calculate?}

\item{p_val_dir}{Direction for p-value calculation, if supplied}
}
\description{
\code{htest_infer_sim} works through the Infer package to  generate samples and a p-value from a 'null hypothesis distribution'. Written (mainly) by Oska Fentem, originally called "test_hypothesis"
}
\details{
' Given a formula, null hypothesis and arguments for generating samples:
Generate samples under this null
#' If p_val_dir is supplied, then the function computes a p-value for observed data given the null
}
\examples{

mtcarsX <- mtcars \%>\% mutate(gear_d_carb = gear - carb)
htest_infer_sim(df = mtcarsX,
stat = "median",
hypothesis = list(null = "point", med = 0),
gen =  list(reps = 1000, type = "bootstrap"),
response = "gear_d_carb",
p_val_dir = "two_sided"))
}
