% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxopts.R
\name{huxoptions}
\alias{huxoptions}
\title{Functions to set some options that look OK when using the huxtable package}
\usage{
huxoptions(df)
}
\description{
\code{huxoptions} for any huxtable tidy table object thing
\code{huxreg_opts} for any huxtable regression tan;es
}
\details{
used in engagement_20

requires dplyr and huxtable packages
}
