% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.med.R
\name{impute.med}
\alias{impute.med}
\title{Replaces missing values with medians for column (ignoring NA's in computing median)}
\usage{
impute.med(x)
}
\arguments{
\item{x}{column of numeric values}
}
\description{
Replaces missing values with medians for column (ignoring NA's in computing median)
}
\examples{
x <- c(0:10, 100, rep(NA,10), 1000, NA)
impute.med(x)

}
