% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_str_lists.R
\name{lab_list_to_text}
\alias{lab_list_to_text}
\title{Functions to deal with lists and text}
\usage{
lab_list_to_text(df)
}
\description{
\code{lab_list_to_text} turns a labled list into text, mainly used for inline code
}
\details{
requires labelled package
}
\examples{
 Demographics: `r eas_all_s \%>\% select(all_of(key_demog)) \%>\% lab_list_to_text`
}
