% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_formulas.R
\name{make_formula}
\alias{make_formula}
\title{Making formulas
\code{m_f} simply collapses lists into a formula notation and makes it a formula
\code{make_formula} does something similar but in a more sophisticated way. using the 'reformulate' function. We use this in predictive modeling
\code{make_formula_df} attatches thi formula to a tibble (multiple) data frames?
\code{wdwd} is a function that...}
\usage{
make_formula(lhs, rhs)
}
\description{
Making formulas
\code{m_f} simply collapses lists into a formula notation and makes it a formula
\code{make_formula} does something similar but in a more sophisticated way. using the 'reformulate' function. We use this in predictive modeling
\code{make_formula_df} attatches thi formula to a tibble (multiple) data frames?
\code{wdwd} is a function that...
}
\details{
Requires stats package
}
\examples{
t0_satis_vars <- c("age_approx", "d_male", "d_student", "not_just_white",  "d_live_usa", "years_involved_norm")
sat_outcome <- "as.factor(satisfaction_ideal)"
eas_20 \%>\%  MASS::polr(m_f(sat_outcome, t0_satis_vars), data = ., Hess=TRUE)

control_vars <- c("ln_years_involved", "year_f", "age"
showup_f <- make_formula("showup", control_vars)

}
