% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_df.R
\name{make_model_df}
\alias{make_model_df}
\title{make a dataframe for modeling
\code{make_model_df} makes a dataframe with outcome variable, model formula and associated dataframe}
\usage{
make_model_df(indep_vars, outcome_vars, dfs)
}
\description{
make a dataframe for modeling
\code{make_model_df} makes a dataframe with outcome variable, model formula and associated dataframe
}
\details{
(uses make_formula (which we just moved to the rethinkpriorities package)
}
\examples{
qp_models <- make_model_df(qp_rhs_vars, qp_outcome_vars, qp_dfs) # Create dataframe for models
}
