% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn_med_sd_na_rm.R
\name{mn}
\alias{mn}
\title{Some shortcut functions for mean, median, and sd,  removing missing values}
\usage{
mn(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
\code{mn} calculates the mean, removing missings
\code{med} calculates the mean, removing missings
\code{sdev} calculates the standard deviation, removing missings
}
\examples{

x <- c(0:10, 100, rep(NA,10))
mn(x)
med(x)
sdev(x)
}
