% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_formatting.R
\name{op}
\alias{op}
\title{Number formatting options, mainly useful for inline code}
\usage{
op(x, d = 3)
}
\arguments{
\item{x}{is something that produces a number}
}
\description{
\code{op} default to 3-digit 'f' formatting with a comma, avoiding scientific notation
}
\examples{
 we see that `r op(share_future_if_email*100)`\% appear in a future EA survey.
}
