% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_formatting.R
\name{ops}
\alias{ops}
\title{\code{ops} limits to 2 digits after decimal point}
\usage{
ops(x, d = 3, ns = 2)
}
\description{
\code{ops} limits to 2 digits after decimal point
}
\examples{
 the median percentage of income donated was `r ops(med_don_share_imp_bc*100)`\% for `r year_n-1`.
}
