% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_cbk.R
\name{rdr_cbk}
\alias{rdr_cbk}
\title{Convenience function to make codebooks with options, particular file structure}
\usage{
rdr_cbk(cbfile)
}
\arguments{
\item{cbfile}{name of 'codebook Rmd' being built}
}
\description{
\code{wdwd} is a function that...
}
\details{
Be careful -- this only works if you have docs and codebooks folders in the right places

requires codebook package to be installed
}
\examples{
rdr_cbk("codebook_eas_all.Rmd")
}
