% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_from_repo.R
\name{read_file_from_repo}
\alias{read_file_from_repo}
\title{Read a file from a (personal) GitHub repository}
\usage{
read_file_from_repo(
  repo,
  path,
  token_key,
  user = "rethinkpriorities",
  private = FALSE
)
}
\arguments{
\item{repo}{A character string specifying the GitHub repository.}

\item{path}{A character string specifying the file path in the GitHub
repository.}

\item{token_key}{A character string specifying the key to the GitHub personal
access token (PAT), stored using the \code(keyring) package.}

\item{user}{A character string specifying the GitHub user name. By default
this is 'rethinkpriorities'.}

\item{private}{A boolean value indicating whether the GitHub repository is a
private repository or not. The default is FALSE.}
}
\description{
\code{read_file_from_repo} reads a file from a GitHub repository.
}
