% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_str_lists.R
\name{remove_str_list}
\alias{remove_str_list}
\title{\code{remove_str_list} -- not sure what it does}
\usage{
remove_str_list(list, string)
}
\description{
\code{remove_str_list} -- not sure what it does
}
