% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_to_varlabel.R
\name{rename_to_var_label}
\alias{rename_to_var_label}
\title{Rename a variable with it's label}
\usage{
rename_to_var_label(df)
}
\arguments{
\item{df}{}
}
\description{
Rename a variable with it's label
}
\details{
We haven't used this on EAS (yet) but it seems so useful to me that I think we should keep it in our package
}
\examples{
p_load(expss)
mtcars <- apply_labels(mtcars[,1:3] ,
                      mpg = "Miles/(US) gallon",
                      cyl = "Number of cylinders",
                      disp = "Displacement (cu.in.)",
                     )
rename_to_var_label(mtcars)
}
