% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_bin.R
\name{se_bin}
\alias{se_bin}
\title{Calculate the SE of a bin}
\usage{
se_bin(x)
}
\arguments{
\item{x}{A vector of binary values that is numeric or can be coerced to be numeric.}
}
\description{
\code{se_bin} calculates the SE of input vector
}
\examples{
# Here is an example
x <- sample(x = c(0, 1),  size = 10, replace = TRUE)
se_bin(x)

}
