% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_det.R
\name{str_det}
\alias{str_det}
\title{Case insensitive string_detect}
\usage{
str_det(string, pattern, negate = FALSE)
}
\description{
Case insensitive string_detect
}
\examples{
num_savdon <- sum(str_det(eas_20$donate_later, "Yes"))
}
