% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumtab_se.R
\name{sumtab_se}
\alias{sumtab_se}
\title{A styled summary table for one outcome by another grouping variable. A lovely set of statistics. Now eith se}
\usage{
sumtab_se(df, depvar, treatvar, caption = "", digits = 3, label = TRUE)
}
\description{
A styled summary table for one outcome by another grouping variable. A lovely set of statistics. Now eith se
}
\details{
uses kable and kable_styling
}
\examples{
 sumtab_se(movies_long, budget, genre)
 airquality \%>\% mutate(ifelse(Ozonesumtab(airquality, Ozone, Month)
}
