% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabylme.R
\name{tabyl_ow}
\alias{tabyl_ow}
\title{simple tables and tabsum ####}
\usage{
tabyl_ow(df, var)
}
\arguments{
\item{var}{is the column to tabulate}

\item{caption}{can contain text}
}
\description{
\code{tabyl_ow}  makes a one-way 'tabyl' style table, arranged in descending order of frequency, not showing missing levels, and making it actually percents
\code{tabyl_ow_plus} does the same as tabyl_ow, but also kables and styles it
\code{tabylme} is a function that makes a two-way 'tabyl' style table, adding row percentages, and kabling and styling it
}
\details{
You need the janitor package, dplyr, and kableExtra
}
\examples{
 tabyl_ow_plus(iris, Species, caption = "iris species or something")
}
