% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabylme.R
\name{tabyl_ow_plus}
\alias{tabyl_ow_plus}
\title{simple tables and tabsum ####}
\usage{
tabyl_ow_plus(df, var, caption = NULL, title_case = FALSE)
}
\arguments{
\item{var}{is the column to tabulate}

\item{caption}{can contain text}
}
\description{
\code{tabyl_ow_plus} is a function that makes a one-way 'tabyl' style table, arranged in descending order of frequency,  and kabling and styling it
\code{tabylme} is a function that makes a two-way 'tabyl' style table, adding row percentages, and kabling and styling it
}
\details{
You need the janitor package, dplyr, and kableExtra
}
\examples{
 tabyl_ow_plus(iris, Species, caption = "iris species or something")
}
