% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_table.R
\name{top_table}
\alias{top_table}
\title{top_table takes the name (outcol) of the top value of outvar in a tidy table, bot_table takes the bottom value}
\usage{
top_table(df, outcol, outvar)
}
\arguments{
\item{outcol}{Is the name of the row of interest}

\item{outvar}{Is the column we want the top value of}
}
\description{
top_table takes the name (outcol) of the top value of outvar in a tidy table, bot_table takes the bottom value
}
\details{
When you make a table using tabyl etc., you often want to find the 'top value row' to use in inline code
}
\examples{

 video_outcomes_phil <- gg_video_breakdowns \%>\%
    filter(audience=="Philanthropy") \%>\%
    group_by(video_theme) \%>\%
    base_results_sum() 

top_outcome <- video_outcomes_phil \%>\% top_table(video_theme, `$/ result`)
}
