% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_to_fn.R
\name{tufte_to_footnote}
\alias{tufte_to_footnote}
\title{Parse a text file; Convert Reinstein's adaptation of tufte notes into standard markdown footnotes}
\usage{
tufte_to_footnote(filename)
}
\description{
\code{wdwd} is a function that...
}
\examples{
ch_Rmds <- purrr::map(ch_Rmd, tufte_to_footnote)

}
