% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model_statement.R
\name{write_model_statement}
\alias{write_model_statement}
\title{Create and write a model statement to a file}
\usage{
write_model_statement(x, file)
}
\description{
This function takes a lavaan model object, converts it to a model
statement, and saves the result to a text file. This text file can be used
to calculate dynamic fit indices via https://www.dynamicfit.app/connect/.
}
