% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_to_missing.R
\name{zero_to_missing}
\alias{zero_to_missing}
\title{Missing to zero and vice versa}
\usage{
zero_to_missing(x)
}
\arguments{
\item{x}{what inputs to function}
}
\description{
\code{zero_to_missing} turns 0 values into NA values in a data frame or vector
\code{missing_to_zero} turns NA values into 0 values in a data frame or vector
}
\examples{
zero_to_missing(mtcars)
}
