% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.r
\name{book}
\alias{book}
\title{The Rethink Priorities bookdown output format}
\usage{
book(
  fig_caption = TRUE,
  number_sections = TRUE,
  anchor_sections = TRUE,
  lib_dir = "libs",
  pandoc_args = NULL,
  extra_dependencies = NULL,
  ...
)
}
\description{
The Rethink Priorities bookdown output format
}
