% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_bin.R
\name{se_bin}
\alias{se_bin}
\title{Calculate the SE of a bin}
\usage{
se_bin(x)
}
\arguments{
\item{x}{A vector that bla bla.}
}
\description{
\code{se_bin} calculates the SE of bla bla.
}
\examples{
# Here is an example
x <- c(0:10, 50)
se_bin(x)

}
